[verified] Extract mining subsystem from tradeskills
